+++ gravitee-am-common/src/main/java/io/gravitee/am/common/policy/ExtensionPoint.java
@@ -41,5 +41,7 @@ public enum ExtensionPoint {
     PRE_CONNECT,
     POST_CONNECT,
     PRE_WEBAUTHN_REGISTER,
-    POST_WEBAUTHN_REGISTER
+    POST_WEBAUTHN_REGISTER,
+    PRE_MFA_ENROLLMENT,
+    POST_MFA_ENROLLMENT
 }
