+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-common/src/main/java/io/gravitee/am/gateway/handler/common/flow/impl/FlowManagerImpl.java
@@ -75,7 +75,8 @@ public class FlowManagerImpl extends AbstractService implements FlowManager, Ini
                 Type.RESET_PASSWORD, List.of(PRE_RESET_PASSWORD, POST_RESET_PASSWORD),
                 Type.REGISTRATION_CONFIRMATION, List.of(PRE_REGISTRATION_CONFIRMATION, POST_REGISTRATION_CONFIRMATION),
                 Type.TOKEN, List.of(PRE_TOKEN, POST_TOKEN),
-                Type.WEBAUTHN_REGISTER, List.of(PRE_WEBAUTHN_REGISTER, POST_WEBAUTHN_REGISTER)
+                Type.WEBAUTHN_REGISTER, List.of(PRE_WEBAUTHN_REGISTER, POST_WEBAUTHN_REGISTER),
+                Type.MFA_ENROLLMENT, List.of(PRE_MFA_ENROLLMENT, POST_MFA_ENROLLMENT)
         );
     }
 
@@ -215,48 +216,49 @@ public class FlowManagerImpl extends AbstractService implements FlowManager, Ini
         var postPolicies = loadPolicies(flow.getPost());
 
         switch (flow.getType()) {
-            case ROOT:
-                // for root type, fetch only the pre step policies
+            case ROOT ->
                 addExecutionFlow(ROOT, flow, prePolicies);
-                break;
-            case CONSENT:
+            case CONSENT -> {
                 addExecutionFlow(PRE_CONSENT, flow, prePolicies);
                 addExecutionFlow(POST_CONSENT, flow, postPolicies);
-                break;
-            case LOGIN:
+            }
+            case LOGIN -> {
                 addExecutionFlow(PRE_LOGIN, flow, prePolicies);
                 addExecutionFlow(POST_LOGIN, flow, postPolicies);
-                break;
-            case LOGIN_IDENTIFIER:
+            }
+            case LOGIN_IDENTIFIER -> {
                 addExecutionFlow(PRE_LOGIN_IDENTIFIER, flow, prePolicies);
                 addExecutionFlow(POST_LOGIN_IDENTIFIER, flow, postPolicies);
-                break;
-            case REGISTER:
+            }
+            case REGISTER -> {
                 addExecutionFlow(PRE_REGISTER, flow, prePolicies);
                 addExecutionFlow(POST_REGISTER, flow, postPolicies);
-                break;
-            case RESET_PASSWORD:
+            }
+            case RESET_PASSWORD -> {
                 addExecutionFlow(PRE_RESET_PASSWORD, flow, prePolicies);
                 addExecutionFlow(POST_RESET_PASSWORD, flow, postPolicies);
-                break;
-            case REGISTRATION_CONFIRMATION:
+            }
+            case REGISTRATION_CONFIRMATION -> {
                 addExecutionFlow(PRE_REGISTRATION_CONFIRMATION, flow, prePolicies);
                 addExecutionFlow(POST_REGISTRATION_CONFIRMATION, flow, postPolicies);
-                break;
-            case TOKEN:
+            }
+            case TOKEN -> {
                 addExecutionFlow(PRE_TOKEN, flow, prePolicies);
                 addExecutionFlow(POST_TOKEN, flow, postPolicies);
-                break;
-            case CONNECT:
+            }
+            case CONNECT -> {
                 addExecutionFlow(PRE_CONNECT, flow, prePolicies);
                 addExecutionFlow(POST_CONNECT, flow, postPolicies);
-                break;
-            case WEBAUTHN_REGISTER:
+            }
+            case WEBAUTHN_REGISTER -> {
                 addExecutionFlow(PRE_WEBAUTHN_REGISTER, flow, prePolicies);
                 addExecutionFlow(POST_WEBAUTHN_REGISTER, flow, postPolicies);
-                break;
-            default:
-                throw new IllegalArgumentException("No suitable flow type found for : " + flow.getType());
+            }
+            case MFA_ENROLLMENT -> {
+                addExecutionFlow(PRE_MFA_ENROLLMENT, flow, prePolicies);
+                addExecutionFlow(POST_MFA_ENROLLMENT, flow, postPolicies);
+            }
+            default -> throw new IllegalArgumentException("No suitable flow type found for : " + flow.getType());
         }
     }
