+++ gravitee-am-model/src/main/java/io/gravitee/am/model/flow/Type.java
@@ -29,5 +29,6 @@ public enum Type {
     RESET_PASSWORD,
     REGISTRATION_CONFIRMATION,
     TOKEN,
-    WEBAUTHN_REGISTER
+    WEBAUTHN_REGISTER,
+    MFA_ENROLLMENT
 }
