+++ gravitee-am-service/src/test/java/io/gravitee/am/service/FlowServiceTest.java
@@ -87,7 +87,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(10);
+        testObserver.assertValueCount(11);
     }
 
     @Test
@@ -97,7 +97,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(10);
+        testObserver.assertValueCount(11);
     }
 
     @Test
@@ -109,7 +109,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(10);
+        testObserver.assertValueCount(11);
     }
 
     @Test
@@ -129,7 +129,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(11);
+        testObserver.assertValueCount(12);
     }
 
     @Test
@@ -139,7 +139,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(10);
+        testObserver.assertValueCount(11);
     }
 
     @Test
@@ -163,7 +163,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(12);
+        testObserver.assertValueCount(13);
     }
 
     @Test
@@ -186,7 +186,7 @@ public class FlowServiceTest {
         testObserver.awaitDone(10, TimeUnit.SECONDS);
         testObserver.assertComplete();
         testObserver.assertNoErrors();
-        testObserver.assertValueCount(11);
+        testObserver.assertValueCount(12);
     }
 
     @Test
